Reorganize packages for test classes.
