+++ app/src/androidTest/java/com/github/leuludyha/ibdb/presentation/components/search/BookSearchTest.kt
@@ -1,4 +1,4 @@
-package com.github.leuludyha.ibdb
+package com.github.leuludyha.ibdb.presentation.components.search
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -15,8 +15,6 @@ import androidx.test.rule.GrantPermissionRule
 import com.github.leuludyha.domain.model.library.*
 import com.github.leuludyha.domain.repository.LibraryRepository
 import com.github.leuludyha.domain.useCase.SearchRemotelyUseCase
-import com.github.leuludyha.ibdb.presentation.components.search.BookSearch
-import com.github.leuludyha.ibdb.presentation.components.search.BookSearchViewModel
 import com.github.leuludyha.ibdb.presentation.navigation.Screen
 import com.github.leuludyha.ibdb.presentation.screen.search.barcode.BarcodeScreen
 import com.github.leuludyha.ibdb.presentation.screen.search.barcode.BarcodeScreenViewModel
